feat: Establish production-ready backend with Python 3.11 dependencies
- Upgrade to Pydantic v2 with compatible FastAPI (0.109.0)
- Update all 15 dependencies for Python 3.11 compatibility
- Fix database schema alignment (credits column)
- Validate complete pipeline: reset → import → generate
- Backend ready for feature development
