+++ backend/alembic/versions/5da0737eb3f8_initial_schema_with_credits.py
@@ -1,0 +1,26 @@
+"""Initial schema with credits
+
+Revision ID: 5da0737eb3f8
+Revises: 
+Create Date: 2026-01-24 23:17:54.902761
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '5da0737eb3f8'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
