--- backend/alembic/versions/e55b5c69ce00_initial_schema.py
@@ -1,100 +1,0 @@
-"""Initial schema
-
-Revision ID: e55b5c69ce00
-Revises: 
-Create Date: 2026-01-22 23:13:39.089915
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e55b5c69ce00'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('assignments', 'room_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.alter_column('assignments', 'timeslot_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.add_column('courses', sa.Column('type', sa.String(), nullable=False, server_default='LECTURE'))
-    op.add_column('courses', sa.Column('weekly_periods', sa.Integer(), nullable=False, server_default='3'))
-    op.add_column('courses', sa.Column('needs_room_type', sa.String(), nullable=False, server_default='LECTURE'))
-    op.drop_column('courses', 'credits')
-    op.add_column('faculty', sa.Column('code', sa.String(), nullable=False, server_default=sa.text("md5(random()::text)")))
-    op.alter_column('faculty', 'email',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.create_unique_constraint(None, 'faculty', ['code'])
-    op.add_column('rooms', sa.Column('code', sa.String(), nullable=False, server_default=sa.text("md5(random()::text)")))
-    op.add_column('rooms', sa.Column('block', sa.String(), nullable=True))
-    op.add_column('rooms', sa.Column('room_no', sa.String(), nullable=True))
-    op.add_column('rooms', sa.Column('capacity', sa.Integer(), nullable=False, server_default='30'))
-    op.drop_constraint('rooms_name_key', 'rooms', type_='unique')
-    op.create_unique_constraint(None, 'rooms', ['code'])
-    op.drop_column('rooms', 'name')
-    op.add_column('sections', sa.Column('code', sa.String(), nullable=False, server_default=sa.text("md5(random()::text)")))
-    op.add_column('sections', sa.Column('dept', sa.String(), nullable=True))
-    op.add_column('sections', sa.Column('program', sa.String(), nullable=True))
-    op.add_column('sections', sa.Column('year', sa.Integer(), nullable=False, server_default='1'))
-    op.add_column('sections', sa.Column('sem', sa.String(), nullable=True))
-    op.alter_column('sections', 'shift',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.create_unique_constraint(None, 'sections', ['code'])
-    op.drop_constraint('sections_course_id_fkey', 'sections', type_='foreignkey')
-    op.drop_constraint('sections_faculty_id_fkey', 'sections', type_='foreignkey')
-    op.drop_column('sections', 'course_id')
-    op.drop_column('sections', 'faculty_id')
-    op.drop_column('sections', 'room_type')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('sections', sa.Column('room_type', sa.VARCHAR(), autoincrement=False, nullable=False))
-    op.add_column('sections', sa.Column('faculty_id', sa.INTEGER(), autoincrement=False, nullable=True))
-    op.add_column('sections', sa.Column('course_id', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.create_foreign_key('sections_faculty_id_fkey', 'sections', 'faculty', ['faculty_id'], ['id'])
-    op.create_foreign_key('sections_course_id_fkey', 'sections', 'courses', ['course_id'], ['id'])
-    op.drop_constraint(None, 'sections', type_='unique')
-    op.alter_column('sections', 'shift',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.drop_column('sections', 'sem')
-    op.drop_column('sections', 'year')
-    op.drop_column('sections', 'program')
-    op.drop_column('sections', 'dept')
-    op.drop_column('sections', 'code')
-    op.add_column('rooms', sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'rooms', type_='unique')
-    op.create_unique_constraint('rooms_name_key', 'rooms', ['name'])
-    op.drop_column('rooms', 'capacity')
-    op.drop_column('rooms', 'room_no')
-    op.drop_column('rooms', 'block')
-    op.drop_column('rooms', 'code')
-    op.drop_constraint(None, 'faculty', type_='unique')
-    op.alter_column('faculty', 'email',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.drop_column('faculty', 'code')
-    op.add_column('courses', sa.Column('credits', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_column('courses', 'needs_room_type')
-    op.drop_column('courses', 'weekly_periods')
-    op.drop_column('courses', 'type')
-    op.alter_column('assignments', 'timeslot_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    op.alter_column('assignments', 'room_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    # ### end Alembic commands ###
